fix: Both mutation and sv/fusion data can be submitted in one ui to create custom mds3 tk
