+++ release.txt
@@ -1,2 +1,5 @@
 Features:
 - enable downloading data for oncoMatrix
+
+Fixes:
+- Both mutation and sv/fusion data can be submitted in one ui to create custom mds3 tk
